gcs_to_bq_pubsub/requirements.txt
google-cloud-bigquery>=1.2.0
google-cloud-storage>=1.6.0
google-cloud-pubsub>=0.39.1
